fix warning: cast GL_MAX_SAMPLES and value-init lights vector
- renderer.hpp: cast glGetIntegerv result to std::uint32_t to silence
  signed/unsigned comparison with sample_count.
- render_context.hpp: value-initialize lights vector to silence
  uninitialized-field warning.

+++ impl/opengl/include/axgl/impl/opengl/renderer.hpp
@@ -66,7 +66,7 @@ public:
     GLint max_samples;
     glGetIntegerv(GL_MAX_SAMPLES, &max_samples);
 
-    if (sample_count > max_samples)
+    if (sample_count > static_cast<std::uint32_t>(max_samples))
     {
       AXGL_LOG_WARN("GL_MAX_SAMPLES: {}", max_samples);
       sample_count_ = max_samples;
